Move auth login to simpleauth Co-authored-by: roesnerb <roesnerb@users.noreply.github.com>
